added confluent-kafka to the website, messages are now send to a topic
