+++ confluent-kafka/schema-registry.properties
@@ -1,0 +1,4 @@
+host.name=schema-registry
+schema.compatibility.level=none
+kafkastore.bootstrap.servers=broker:29092
+listeners=http://0.0.0.0:8081
