+++ python-app/app.py
@@ -1,5 +1,10 @@
 from flask import Flask, request, render_template, jsonify
+from confluent import Demonstator
+
 app = Flask(__name__)
+
+# create a confluent-kafka instance
+my_messages = Demonstator("172.29.0.3","9092","172.29.0.4","8081")
 
 @app.route('/')
 def index():
@@ -16,6 +21,7 @@ def login():
     user = request.form['name']
 
     # Kafka goes here
+    my_messages.produceMessage("user",{"ID":123,"username":user},{"ID":123,"username":user})
 
     return render_template('user.html', user=user)
 
@@ -26,8 +32,10 @@ def send():
     message = request.args.get('message', 'None', type=str)
 
     # Kafka goes here
+    my_messages.produceMessage("user",{"ID":123,"username":message},{"ID":123,"username":message})
+
 
     return jsonify(last_message=message)
 
 if __name__ == '__main__':
-    app.run()
+    app.run(host="0.0.0.0", port=5000,)
